+++ SY/task/task5a.php
@@ -2,7 +2,7 @@
 
 echo 'Hi,me again ! Now i can help you convert MYR to USD dollar ! ';
 
-$MYR_input = (int)readline('So how much MYR you have ?' . PHP_EOL);
+$MYR_input = (float)readline('So how much MYR you have ?' . PHP_EOL);
 
 $USD = ($MYR_input / 4.19);
